feat: agregada creacion de tareas junto con la subidad de ficheros a aws

+++ src/main/java/com/proyect/CodeShareSpace/dto/task/FileTasksDto.java
@@ -12,6 +12,6 @@ public class FileTasksDto {
     private String fileName;
     private String prefix;
     private String type;
-    private Double KBytes;
+    private Long KBytes;
     private Long taskId;
 }
